Warning cleanup
git-svn-id: http://svn.l2jfree.com/l2jfree@5171 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-geoserver/src/main/java/com/l2jfree/geoserver/util/Util.java
@@ -46,10 +46,10 @@ public final class Util
 
 	public static String[] getMemUsage()
 	{
-		double maxMem = ((long) (Runtime.getRuntime().maxMemory() / 1024)); // maxMemory is the upper limit the jvm can use
-		double allocatedMem = ((long) (Runtime.getRuntime().totalMemory() / 1024)); // totalMemory the size of the current allocation pool
+		double maxMem = Runtime.getRuntime().maxMemory() / 1024; // maxMemory is the upper limit the jvm can use
+		double allocatedMem = Runtime.getRuntime().totalMemory() / 1024; // totalMemory the size of the current allocation pool
 		double nonAllocatedMem = maxMem - allocatedMem; // non allocated memory till jvm limit
-		double cachedMem = ((long) (Runtime.getRuntime().freeMemory() / 1024)); // freeMemory the unused memory in the allocation pool
+		double cachedMem = Runtime.getRuntime().freeMemory() / 1024; // freeMemory the unused memory in the allocation pool
 		double usedMem = allocatedMem - cachedMem; // really used memory
 		double useableMem = maxMem - usedMem; // allocated, but non-used and non-allocated memory
 		return new String[]
